chore: fix blog card component
