--- components/Blog/Card/Img.tsx
@@ -1,9 +1,0 @@
-import Image, { ImageProps } from 'next/image';
-
-interface Props extends ImageProps { }
-
-const Img = ({ ...props }: Props) => {
-   return <Image {...props} />;
-};
-
-export default Img;
